861m9x2bf: controls configurator in separate components
